fix: loading 관련 아이콘 수정 및 새로 생성
