+++ components/atoms/icons/loading-icon.tsx
@@ -1,43 +1,17 @@
-import { css } from '@/styled-system/css';
-
-const spinAnimation = css({
-  animation: 'spin 1s linear infinite',
-});
-
-export const LoadingIcon = () => {
+export const LoadingIcon = (props?: React.SVGProps<SVGSVGElement>) => {
   return (
     <svg
-      width="21"
-      height="20"
-      viewBox="0 0 21 20"
+      width="407"
+      height="409"
+      viewBox="0 0 407 409"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
-      className={spinAnimation}
+      {...props}
     >
-      <mask
-        id="mask0_18350_10256"
-        style={{ maskType: 'alpha' }}
-        maskUnits="userSpaceOnUse"
-        x="0"
-        y="0"
-        width="21"
-        height="20"
-      >
-        <path
-          d="M16.8841 4.64328C18.3047 6.33634 18.9946 8.52441 18.802 10.7261C18.6094 12.9279 17.55 14.9629 15.8569 16.3835C14.1639 17.8042 11.9758 18.4941 9.77405 18.3015C7.57232 18.1088 5.53728 17.0495 4.11663 15.3564L5.39522 14.2835C6.53133 15.6375 8.15877 16.4847 9.91952 16.6387C11.6803 16.7928 13.4301 16.2411 14.7841 15.105C16.138 13.9688 16.9852 12.3414 17.1393 10.5807C17.2933 8.81992 16.7416 7.0701 15.6055 5.71614L16.8841 4.64328Z"
-          fill="white"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M10.5004 18.3332C15.1027 18.3332 18.8337 14.6022 18.8337 9.99984C18.8337 5.39746 15.1027 1.6665 10.5004 1.6665C5.89797 1.6665 2.16699 5.39746 2.16699 9.99984C2.16699 14.6022 5.89797 18.3332 10.5004 18.3332ZM10.5004 16.6665C14.1823 16.6665 17.167 13.6817 17.167 9.99984C17.167 6.31794 14.1823 3.33317 10.5004 3.33317C6.81844 3.33317 3.83366 6.31794 3.83366 9.99984C3.83366 13.6817 6.81844 16.6665 10.5004 16.6665Z"
-          fill="white"
-          fillOpacity="0.25"
-        />
-      </mask>
-      <g mask="url(#mask0_18350_10256)">
-        <rect x="0.5" width="20" height="20" fill="white" />
-      </g>
+      <path
+        d="M372.327 204.5C392.581 204.5 409.334 187.951 405.721 168.022C399.547 133.968 384.789 101.828 362.581 74.7666C332.566 38.1938 290.799 13.1596 244.396 3.92941C197.993 -5.30073 149.825 1.8443 108.099 24.1471C66.3738 46.4499 33.6722 82.5305 15.5666 126.241C-2.53894 169.952 -4.92829 218.588 8.8057 263.863C22.5397 309.138 51.5473 348.25 90.8859 374.536C130.224 400.821 177.46 412.653 224.544 408.015C259.384 404.584 292.545 392.293 320.991 372.578C337.637 361.041 337.493 337.493 323.171 323.171V323.171C308.85 308.85 285.76 309.421 268.042 319.233C252.509 327.835 235.278 333.257 217.355 335.022C187.158 337.996 156.864 330.408 131.635 313.55C106.406 296.693 87.8019 271.608 78.9938 242.572C70.1857 213.535 71.7181 182.343 83.3299 154.31C94.9417 126.276 115.914 103.136 142.675 88.8328C169.435 74.5292 200.327 69.9468 230.087 75.8665C259.847 81.7861 286.634 97.8415 305.883 121.297C317.308 135.219 325.659 151.236 330.56 168.302C336.15 187.769 352.073 204.5 372.327 204.5V204.5Z"
+        fill="#3B87F4"
+      />
     </svg>
   );
 };
